+++ build.ps1
@@ -18,10 +18,10 @@ $ErrorActionPreference = "Stop"
 
 Import-module "$PSScriptRoot/.psscripts/build-functions.ps1" -Force
 
-Write-BuildHeader "Starting giraffe-template build script"
+Write-BuildHeader "Starting giraffe-api-template build script"
 
-$nuspec = "./src/giraffe-template.nuspec"
-$version = Get-NuspecVersion $nuspec
+$buildfile = "./src/giraffe-api-template.csproj"
+$version = Get-BuildVersion $buildfile
 
 Update-AppVeyorBuildVersion $version
 
@@ -45,43 +45,10 @@ dotnet-restore $giraffeTests
 dotnet-build   $giraffeTests
 dotnet-test    $giraffeTests
 
-# Test Razor template
-Write-Host "Building and testing Razor tempalte..." -ForegroundColor Magenta
-$razorApp       = "src/content/Razor/src/AppNamePlaceholder/AppNamePlaceholder.fsproj"
-$razorTests     = "src/content/Razor/tests/AppNamePlaceholder.Tests/AppNamePlaceholder.Tests.fsproj"
-
-dotnet-restore $razorApp
-dotnet-build   $razorApp
-dotnet-restore $razorTests
-dotnet-build   $razorTests
-dotnet-test    $razorTests
-
-# Test DotLiquid template
-Write-Host "Building and testing DotLiquid tempalte..." -ForegroundColor Magenta
-$dotLiquidApp   = "src/content/DotLiquid/src/AppNamePlaceholder/AppNamePlaceholder.fsproj"
-$dotLiquidTests = "src/content/DotLiquid/tests/AppNamePlaceholder.Tests/AppNamePlaceholder.Tests.fsproj"
-
-dotnet-restore $dotLiquidApp
-dotnet-build   $dotLiquidApp
-dotnet-restore $dotLiquidTests
-dotnet-build   $dotLiquidTests
-dotnet-test    $dotLiquidTests
-
-# Test None template
-Write-Host "Building and testing None tempalte..." -ForegroundColor Magenta
-$noneApp   = "src/content/None/src/AppNamePlaceholder/AppNamePlaceholder.fsproj"
-$noneTests = "src/content/None/tests/AppNamePlaceholder.Tests/AppNamePlaceholder.Tests.fsproj"
-
-dotnet-restore $noneApp
-dotnet-build   $noneApp
-dotnet-restore $noneTests
-dotnet-build   $noneTests
-dotnet-test    $noneTests
-
 # Create template NuGet package
 Remove-OldBuildArtifacts
 Write-Host "Building NuGet package..." -ForegroundColor Magenta
-Invoke-Cmd "nuget pack src/giraffe-template.nuspec"
+Invoke-Cmd "dotnet pack src/giraffe-api-template.csproj"
 
 if ($UpdatePaketDependencies.IsPresent -or $TestPermutations.IsPresent -or $CreatePermutations.IsPresent -or $InstallTemplate.IsPresent)
 {
@@ -89,17 +56,16 @@ if ($UpdatePaketDependencies.IsPresent -or $TestPermutations.IsPresent -or $Crea
     Write-Host "Uninstalling existing Giraffe template..." -ForegroundColor Magenta
     $giraffeInstallation = Invoke-UnsafeCmd "dotnet new giraffe --list"
     $giraffeInstallation
-    if ($giraffeInstallation[$giraffeInstallation.Length - 2].StartsWith("Giraffe Web App"))
+    if ($giraffeInstallation[$giraffeInstallation.Length - 2].StartsWith("Giraffe API"))
     {
-        Invoke-Cmd "dotnet new -u giraffe-template"
+        Invoke-Cmd "dotnet new -u giraffe-api-template"
     }
-    # if ($giraffeInstallation.Length -lt 6) { Invoke-Cmd "dotnet new -u giraffe-template" }
 
-    $nupkg     = Get-ChildItem "./giraffe-template.$version.nupkg"
+    $nupkg     = Get-ChildItem "./src/bin/Debug/giraffe-api-template.$version.nupkg"
     $nupkgPath = $nupkg.FullName
 
     # Installing Giraffe template
-    Write-Host "Installing newly built Giraffe tempalte..." -ForegroundColor Magenta
+    Write-Host "Installing newly built Giraffe template..." -ForegroundColor Magenta
     Invoke-Cmd "dotnet new -i $nupkgPath"
 
     if ($UpdatePaketDependencies.IsPresent -or $TestPermutations.IsPresent -or $CreatePermutations.IsPresent)
@@ -110,27 +76,21 @@ if ($UpdatePaketDependencies.IsPresent -or $TestPermutations.IsPresent -or $Crea
         if (Test-Path $tempFolder) { Remove-Item $tempFolder -Recurse -Force }
         New-Item -Name ".temp" -ItemType Directory
 
-        # Creating all permutations
-        Write-Host "Creating all permutations of the giraffe-tempalte..." -ForegroundColor Magenta
+        $viewEngine = "Giraffe"
+        Write-Host "Creating templates for: $viewEngine..." -ForegroundColor Magenta
 
-        $viewEngines = "Giraffe", "Razor", "DotLiquid", "None"
-        foreach ($viewEngine in $viewEngines)
-        {
-            Write-Host "Creating templates for view engine: $viewEngine..." -ForegroundColor Magenta
+        $engine = $viewEngine.ToLower()
 
-            $engine = $viewEngine.ToLower()
+        Invoke-Cmd ("dotnet new giraffe -lang F# -V $engine -o $tempFolder/$viewEngine" + "App")
+        Invoke-Cmd ("dotnet new giraffe -lang F# -I -V $engine -o $tempFolder/$viewEngine" + "TestsApp")
+        Invoke-Cmd ("dotnet new giraffe -lang F# -U -V $engine -o $tempFolder/$viewEngine" + "PaketApp")
+        Invoke-Cmd ("dotnet new giraffe -lang F# -I -U -V $engine -o $tempFolder/$viewEngine" + "TestsPaketApp")
 
-            Invoke-Cmd ("dotnet new giraffe -lang F# -V $engine -o $tempFolder/$viewEngine" + "App")
-            Invoke-Cmd ("dotnet new giraffe -lang F# -I -V $engine -o $tempFolder/$viewEngine" + "TestsApp")
-            Invoke-Cmd ("dotnet new giraffe -lang F# -U -V $engine -o $tempFolder/$viewEngine" + "PaketApp")
-            Invoke-Cmd ("dotnet new giraffe -lang F# -I -U -V $engine -o $tempFolder/$viewEngine" + "TestsPaketApp")
-        }
 
         if ($UpdatePaketDependencies.IsPresent -or $TestPermutations.IsPresent)
         {
             # Building and testing all permutations
             Write-Host "Building and testing all permutations of the giraffe-tempalte..." -ForegroundColor Magenta
-            $isWin = Test-IsWindows
 
             Get-ChildItem ".temp" -Directory | ForEach-Object {
                 $name = $_.Name
@@ -142,12 +102,7 @@ if ($UpdatePaketDependencies.IsPresent -or $TestPermutations.IsPresent -or $Crea
                     Remove-Item -Path "paket.lock" -Force
                 }
 
-                if ($isWin) {
-                    Invoke-Cmd ("./build.bat")
-                }
-                else {
-                    Invoke-Cmd ("sh ./build.sh")
-                }
+                Invoke-Cmd("sh ./build.sh")
 
                 if ($UpdatePaketDependencies.IsPresent -and $name.Contains("Paket"))
                 {
@@ -161,4 +116,4 @@ if ($UpdatePaketDependencies.IsPresent -or $TestPermutations.IsPresent -or $Crea
     }
 }
 
-Write-SuccessFooter "The giraffe-template has been successfully built!"
+Write-SuccessFooter "The giraffe-api-template has been successfully built!"